refactor(mcp): extract preview tool module
