+++ openspec/changes/refactor-mcp-preview-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-preview-tool-module
+
+Refactor: extract MCP preview tool implementation into a dedicated module
