+++ src/mcp/tools.rs
@@ -17,6 +17,7 @@ use super::{AgentpackMcp, confirm};
 mod evolve_propose;
 mod evolve_restore;
 mod explain;
+mod preview;
 mod rollback;
 mod status;
 
@@ -345,63 +346,7 @@ async fn call_evolve_propose_in_process(
 }
 
 async fn call_preview_in_process(args: PreviewArgs) -> anyhow::Result<(String, serde_json::Value)> {
-    tokio::task::spawn_blocking(move || {
-        let repo_override = args.common.repo.as_ref().map(std::path::PathBuf::from);
-        let profile = args.common.profile.as_deref().unwrap_or("default");
-        let target = args.common.target.as_deref().unwrap_or("all");
-        let machine_override = args.common.machine.as_deref();
-
-        let result = crate::handlers::read_only::read_only_context(
-            repo_override.as_deref(),
-            machine_override,
-            profile,
-            target,
-        );
-
-        let (text, envelope) = match result {
-            Ok(crate::handlers::read_only::ReadOnlyContext {
-                targets,
-                desired,
-                plan,
-                mut warnings,
-                roots,
-            }) => {
-                let data = crate::app::preview_json::preview_json_data(
-                    profile,
-                    targets,
-                    plan,
-                    desired,
-                    roots,
-                    args.diff,
-                    &mut warnings,
-                )?;
-
-                let mut envelope = crate::output::JsonEnvelope::ok("preview", data);
-                envelope.warnings = warnings;
-
-                let text = serde_json::to_string_pretty(&envelope)?;
-                let envelope = serde_json::to_value(&envelope)?;
-                (text, envelope)
-            }
-            Err(err) => {
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                let envelope = envelope_error("preview", &code, &message, details);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                (text, envelope)
-            }
-        };
-
-        Ok((text, envelope))
-    })
-    .await
-    .context("mcp preview handler task join")?
+    preview::call_preview_in_process(args).await
 }
 
 async fn deploy_plan_envelope_in_process(args: CommonArgs) -> anyhow::Result<serde_json::Value> {
