+++ src/mcp/tools/preview.rs
@@ -1,0 +1,65 @@
+use anyhow::Context as _;
+
+use crate::user_error::UserError;
+
+pub(super) async fn call_preview_in_process(
+    args: super::PreviewArgs,
+) -> anyhow::Result<(String, serde_json::Value)> {
+    tokio::task::spawn_blocking(move || {
+        let repo_override = args.common.repo.as_ref().map(std::path::PathBuf::from);
+        let profile = args.common.profile.as_deref().unwrap_or("default");
+        let target = args.common.target.as_deref().unwrap_or("all");
+        let machine_override = args.common.machine.as_deref();
+
+        let result = crate::handlers::read_only::read_only_context(
+            repo_override.as_deref(),
+            machine_override,
+            profile,
+            target,
+        );
+
+        let (text, envelope) = match result {
+            Ok(crate::handlers::read_only::ReadOnlyContext {
+                targets,
+                desired,
+                plan,
+                mut warnings,
+                roots,
+            }) => {
+                let data = crate::app::preview_json::preview_json_data(
+                    profile,
+                    targets,
+                    plan,
+                    desired,
+                    roots,
+                    args.diff,
+                    &mut warnings,
+                )?;
+
+                let mut envelope = crate::output::JsonEnvelope::ok("preview", data);
+                envelope.warnings = warnings;
+
+                let text = serde_json::to_string_pretty(&envelope)?;
+                let envelope = serde_json::to_value(&envelope)?;
+                (text, envelope)
+            }
+            Err(err) => {
+                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
+                let code = user_err
+                    .map(|e| e.code.clone())
+                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
+                let message = user_err
+                    .map(|e| e.message.clone())
+                    .unwrap_or_else(|| err.to_string());
+                let details = user_err.and_then(|e| e.details.clone());
+                let envelope = super::envelope_error("preview", &code, &message, details);
+                let text = serde_json::to_string_pretty(&envelope)?;
+                (text, envelope)
+            }
+        };
+
+        Ok((text, envelope))
+    })
+    .await
+    .context("mcp preview handler task join")?
+}
